requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
playwright==1.47.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
httpx==0.27.2
slack-sdk==3.31.0
jinja2==3.1.4
python-multipart==0.0.9
apscheduler==3.10.4
openpyxl>=3.1.5
pandas>=3.0.0
aiosmtplib==3.0.1
anyio>=4.0.0
starlette>=0.39.0
